requirements.txt
python-telegram-bot
pymupdf
Pillow
requests